reorganized test a bit
(cherry picked from commit 10ec820885ad0f7e02a3abaa133963d98e2bc5b9)
