+++ mimium-lang/tests/intergration_test.rs
@@ -1,36 +1,7 @@
-extern crate mimium_lang;
-use std::{
-    collections::HashMap,
-    path::{Path, PathBuf},
-};
+mod common;
+use common::*;
 
-use mimium_lang::{
-    compiler,
-    runtime::run_source_test,
-    utils::{error::report, fileloader},
-};
 
-fn run_simple_test(expr: &str, expect: f64, times: u64) {
-    let src = format!(
-        "fn test(hoge){{
-    {expr}
-}}
-fn dsp(){{
-    test(2.0)
-}}"
-    );
-    let res = run_source_test(&src, times, false);
-    match res {
-        Ok(res) => {
-            let ans = [expect].repeat(times as usize);
-            assert_eq!(res, ans, "expr: {expr}");
-        }
-        Err(errs) => {
-            report(&src, Path::new("(from template)"), &errs);
-            panic!("invalid syntax");
-        }
-    }
-}
 
 #[test]
 fn simple_arithmetic() {
@@ -54,34 +25,6 @@ fn simple_arithmetic() {
     run_simple_test("1.0+hoge^2.0*1.5", 7.0, 3);
 }
 
-fn run_file_test(path: &str, times: u64, stereo: bool) -> Result<Vec<f64>, ()> {
-    let (file, src) = load_src(path);
-    let res = run_source_test(&src, times, stereo);
-    match res {
-        Ok(res) => Ok(res),
-        Err(errs) => {
-            report(&src, file, &errs);
-            Err(())
-        }
-    }
-}
-
-fn load_src(path: &str) -> (PathBuf, String) {
-    let file: PathBuf = [env!("CARGO_MANIFEST_DIR"), "tests/mmm", path]
-        .iter()
-        .collect();
-    println!("{}", file.to_str().unwrap());
-    let (src, _path) = fileloader::load(file.to_string_lossy().to_string()).unwrap();
-    (file, src)
-}
-
-fn run_file_test_mono(path: &str, times: u64) -> Result<Vec<f64>, ()> {
-    run_file_test(path, times, false)
-}
-
-fn run_file_test_stereo(path: &str, times: u64) -> Result<Vec<f64>, ()> {
-    run_file_test(path, times, true)
-}
 
 #[test]
 fn parser_firstbreak() {
@@ -310,36 +253,6 @@ fn fb_mem2() {
     assert_eq!(res, ans);
 }
 
-fn test_state_sizes<T: IntoIterator<Item = (&'static str, u64)>>(path: &str, ans: T) {
-    let state_sizes: HashMap<&str, u64> = HashMap::from_iter(ans.into_iter());
-    let (file, src) = load_src(path);
-    let bytecode = match compiler::emit_bytecode(&src) {
-        Ok(res) => res,
-        Err(errs) => {
-            report(&src, file, &errs);
-            panic!("failed to emit bytecode");
-        }
-    };
-
-    for (sym, proto) in bytecode.global_fn_table {
-        let fn_name = sym.as_str();
-
-        if fn_name == "_mimium_global" {
-            continue;
-        }
-
-        let actual = proto.state_size;
-        match state_sizes.get(fn_name) {
-            Some(&expected) => {
-                assert_eq!(
-                    actual, expected,
-                    "state size of function `{fn_name}` is wrong"
-                );
-            }
-            None => panic!("no such function: {fn_name}"),
-        };
-    }
-}
 
 #[test]
 fn fb_mem3_state_size() {
